Final clean version with single comprehensive example
